ci/jenkins/pipelines/prs/helpers/pr_manager/requirements.txt
PyGithub==1.43.7
python-jenkins==1.4.0
